Prep: Configure git; Create README.md and License.
To configure git , we add the .gitconfig and .gitignore files.

The .gitconfig file configures the repository. While the file can be
used to configure a number, our project uses the file to provide helper
functions for perusing this repository. Notably, the previous, next, and
chapter commands are defined in the file. Please see the README.md file
for usage information. For more information about .gitconfig, please
see:

    https://git-scm.com/docs/git-config

The .gitignore file tells git which files never to add to version
control. We can use the Python template provided by the gitignore
project:

    https://github.com/github/gitignore

In anticipation of Chapter 3, we tell git to ignore the SQLite database
we will generate. For more information about .gitignore, please see:

    https://git-scm.com/docs/gitignore

Additionally, this commit provides the 2-clause BSD License, as well as
the README file which github.com will use to provide basic project
information.