Ch01: Add organizer and blog apps to settings.
Where a Django project is a website, a Django app is a feature set.

The INSTALLED_APPS setting is an explicit list of all of the features of
our website. Django automatically includes a few; we add our own apps to
the list to add these features to the project (or website).