Ch03: Define Startup model related fields. [skip ci]
    https://docs.djangoproject.com/en/1.8/ref/models/fields/#manytomanyfield

A many-to-many relationship allows for different models to be related
to many of the other. For instance, students may take many classes, and
classes are taught to many students. In our code, we allow for Startups
to be categorized by Tag objects. A Tag object may categorize many
Startup objects, and Startups may be categorized by many Tag objects.