Ch03: Add options to Startup model fields. [skip ci]
Field options allow us to easily customize behavior of a field.

Global Field Options:

    https://docs.djangoproject.com/en/1.8/ref/models/fields/#db-index
    https://docs.djangoproject.com/en/1.8/ref/models/fields/#help-text
    https://docs.djangoproject.com/en/1.8/ref/models/fields/#unique
    https://docs.djangoproject.com/en/1.8/ref/models/fields/#verbose-name

Verbose names are further documented in:

    https://docs.djangoproject.com/en/1.8/topics/db/models/#verbose-field-names

The max_length field option is defined in CharField and inherited by all
CharField subclasses (but is typically optional in these subclasses,
unlike CharField itself).

The 255 character limit of the URLField is based on RFC 3986.

    https://tools.ietf.org/html/rfc3986